+++ src/NumArray.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace NumPHP;
 
+use NumPHP\Exception\MissingArgumentException;
+
 class NumArray
 {
     private $data;
@@ -26,6 +28,11 @@ class NumArray
             $this->string = self::recursiveToString($this->data);
         }
         return $this->string;
+    }
+
+    public function isEqual(NumArray $numArray): bool
+    {
+        return get_class($this) === get_class($numArray) && $this->getData() === $numArray->getData();
     }
 
     public function getData(): array
@@ -69,6 +76,32 @@ class NumArray
         return $this->nDim;
     }
 
+    public static function ones(int ...$axis): NumArray
+    {
+        if (empty($axis)) {
+            throw new MissingArgumentException('No $axis given');
+        }
+        return new NumArray(self::recursiveFillArray($axis, 1));
+    }
+
+    public static function onesLike(NumArray $numArray): NumArray
+    {
+        return call_user_func_array([self, 'ones'], $numArray->getShape());
+    }
+
+    public static function zeros(int ...$axis): NumArray
+    {
+        if (empty($axis)) {
+            throw new MissingArgumentException('No $axis given');
+        }
+        return new NumArray(self::recursiveFillArray($axis, 0));
+    }
+
+    public static function zerosLike(NumArray $numArray): NumArray
+    {
+        return call_user_func_array([self, 'zeros'], $numArray->getShape());
+    }
+
     private static function recursiveToString(array $data, int $level = 0): string
     {
         $indent = str_repeat("  ", $level);
@@ -80,5 +113,13 @@ class NumArray
             return sprintf("%s[\n%s\n%s]", $indent, implode(",\n", $result), $indent);
         }
         return sprintf("%s[%s]", $indent, implode(", ", $data));
+    }
+
+    private static function recursiveFillArray(array $shape, $value): array
+    {
+        if (count($shape) === 1) {
+            return array_fill(0, array_shift($shape), $value);
+        }
+        return array_fill(0, array_shift($shape), self::recursiveFillArray($shape, $value));
     }
 }