+++ CHANGELOG.md
@@ -4,15 +4,20 @@ All Notable changes to `NumPHP` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - 2017-01-31
+## NEXT - 2017-02-01
 
 ### Added
+- Add class `NumPHP\Exception\MissingArgumentException`
 - Add class `NumPHP\NumArray`
  - Add function `__toString`
  - Add function `getData`
  - Add function `getShape`
  - Add function `getSize`
  - Add function `getNDim`
+ - Add fucntion `ones`
+ - Add function `onesLike`
+ - Add function `zeros`
+ - Add function `zerosLike`
 - Add class `NumPHP\ComplexNum`
  - Add function `__toString`
  - Add function `isEqual`